Extract session handling to a module
