+++ include/tlog/Makefile.am
@@ -54,6 +54,7 @@ tlog_HEADERS = \
     rec_session_conf.h          \
     rec_session_conf_cmd.h      \
     rec_session_conf_validate.h \
+    session.h                   \
     sink.h                      \
     sink_type.h                 \
     source.h                    \
